Fix dependency on luckperms, add correct dependencies for some features that need them
