+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/features/AbyssContext.kt
@@ -46,6 +46,7 @@ class AbyssContext(
 
 
     val isChattyLoaded get() = plugin.server.pluginManager.isPluginEnabled("chatty")
+    val isEternalFortuneLoaded get() = plugin.server.pluginManager.isPluginEnabled("EternalFortune")
     val isGSitLoaded get() = plugin.server.pluginManager.isPluginEnabled("GSit")
     val isPlaceholderApiLoaded get() = plugin.server.pluginManager.isPluginEnabled("PlaceholderAPI")
     val isHMCCosmeticsEnabled get() = plugin.server.pluginManager.isPluginEnabled("HMCCosmetics")
