+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/features/guilds/GuildListener.kt
@@ -1,35 +1,18 @@
 package com.mineinabyss.features.guilds
 
 import com.github.shynixn.mccoroutine.bukkit.asyncDispatcher
-import com.mineinabyss.chatty.chatty
-import com.mineinabyss.chatty.components.chattyData
-import com.mineinabyss.chatty.helpers.getDefaultChat
-import com.mineinabyss.chatty.listeners.RendererExtension
 import com.mineinabyss.components.guilds.GuildMaster
-import com.mineinabyss.eternalfortune.api.events.PlayerOpenGraveEvent
 import com.mineinabyss.features.abyss
 import com.mineinabyss.features.guilds.database.GuildMessageQueue
 import com.mineinabyss.features.guilds.database.GuildMessageQueue.content
-import com.mineinabyss.features.guilds.database.GuildRank
-import com.mineinabyss.features.guilds.extensions.getGuildChatId
-import com.mineinabyss.features.guilds.extensions.getGuildName
-import com.mineinabyss.features.guilds.extensions.getGuildRank
-import com.mineinabyss.features.guilds.extensions.hasGuild
 import com.mineinabyss.features.guilds.menus.GuildMainMenu
-import com.mineinabyss.features.helpers.di.Features
-import com.mineinabyss.geary.papermc.tracking.entities.toGeary
 import com.mineinabyss.geary.papermc.tracking.entities.toGearyOrNull
 import com.mineinabyss.guiy.inventory.guiy
 import com.mineinabyss.idofront.messaging.info
 import com.mineinabyss.idofront.textcomponents.miniMsg
-import io.papermc.paper.event.player.AsyncChatEvent
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
-import nl.rutgerkok.blocklocker.group.GroupSystem
-import org.bukkit.Bukkit
-import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
-import org.bukkit.event.EventPriority
 import org.bukkit.event.Listener
 import org.bukkit.event.player.PlayerInteractAtEntityEvent
 import org.bukkit.event.player.PlayerJoinEvent
@@ -56,63 +39,6 @@ class GuildListener : Listener {
                 }
                 GuildMessageQueue.deleteWhere { playerUUID eq player.uniqueId }
             }
-        }
-    }
-
-    @EventHandler
-    fun PlayerOpenGraveEvent.onOpenGuildMemberGrave() {
-        if (grave.isExpired() || !grave.isProtected() || graveOwner.uniqueId == player.uniqueId) return
-
-        isCancelled = when {
-            !player.hasGuild() || !graveOwner.hasGuild() -> true
-            player.getGuildName() != graveOwner.getGuildName() -> true
-            else -> !Features.guilds.config.canOpenGuildMemberGraves
-        }
-    }
-}
-
-class GuildContainerSystem : GroupSystem() {
-
-    override fun isInGroup(player: Player, guildName: String): Boolean {
-        val name = player.getGuildName().replace(" ", "_")
-        return name.equals(guildName, true) && player.hasGuild()
-    }
-
-    override fun isGroupLeader(player: Player, groupName: String): Boolean {
-        val guild = player.hasGuild()
-        if (!guild) return false
-
-        return player.getGuildRank() == GuildRank.OWNER
-    }
-}
-
-class ChattyGuildListener : Listener {
-
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
-    fun AsyncChatEvent.onGuildChat() {
-        val channelId = player.chattyData.channelId
-
-        if ((!channelId.startsWith(player.getGuildName()) && channelId.endsWith(guildChannelId)) || channelId !in chatty.config.channels.keys) {
-            player.toGeary().setPersisting(player.chattyData.copy(channelId = getDefaultChat().key))
-            return
-        }
-        if (player.chattyData.channelId != player.getGuildChatId()) return
-
-        viewers().clear()
-        viewers().addAll(Bukkit.getOnlinePlayers().filter {
-            it.getGuildName() == player.getGuildName() && it != player
-        })
-
-        if (chatty.config.chat.disableChatSigning) {
-            viewers().forEach { a -> RendererExtension.render(player, player.displayName(), message(), a) }
-            viewers().clear()
-        }
-    }
-
-    @EventHandler
-    fun PlayerJoinEvent.onJoin() {
-        if (player.chattyData.channelId.endsWith(guildChannelId)) {
-            player.toGeary().setPersisting(player.chattyData.copy(channelId = player.getGuildChatId().takeIf { it.isNotBlank() } ?: getDefaultChat().key))
         }
     }
 }
